src/pyproject.toml
[project]
name = "nutrisnap"
version = "0.1.0"
description = "NutriSnap - Food Classification with Vision Transformers"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.12.0",
    "evaluate>=0.4.0",
    "wandb>=0.15.0",
    "Pillow>=9.5.0",
    "numpy>=1.24.0",
    "accelerate>=0.20.0",
    "scikit-learn>=1.3.0",
    "gcsfs>=2024.3.1",
]

[tool.uv]